+++ README.md
@@ -122,8 +122,8 @@ Each sample demonstrates one feature of the SDK, together with tests.
   Execution and Activity Executions. Additional
   documentation: [How to use tracing in Go](https://docs.temporal.io/go/tracing).
 
-- [**OpenTelemetry**](./otel): Demonstrates how to instrument the Wokflows and Activities
-  with OpenTelemetry.
+- [**OpenTelemetry**](./opentelemetry): Demonstrates how to instrument the Wokflows and
+  Activities with OpenTelemetry.
 
 - [**Updatable Timer**](./updatabletimer): Demonstrates timer
   cancellation and use of a Selector to wait on a Future and a Channel simultaneously.
